+++ CHANGELOG.md
@@ -6,6 +6,7 @@ the [Semantic Versioning 2.0.0](http://semver.org/) format.
 ### Added
 
 - Make `ErrorDetail` nullable in method parameters
+- `OperateWhen` overloads
 
 ### Updated